refactor(lyrics): consume the published parser package
