+++ src/modules/lyrics/providers/unified.ts
@@ -1,10 +1,9 @@
 import { CUBEY_LYRICS_API_URL, CUBEY_LYRICS_API_URL_TURNSTILE, HOMEPAGE_URL, LOG_PREFIX } from "@constants";
 import { getLocalStorage } from "@core/storage";
 import { log } from "@core/utils";
-import { lrcFixers, parseLRC, parsePlainLyrics } from "./lrcUtils";
-import { parseQRC } from "./qrcUtils";
+import { lrcFixers, parseLRC, parseQRC, PlainParser } from "@braccato/parsers";
 import { type LyricSourceKey, type LyricSourceResult, type ProviderParameters, saveLyricsToCache } from "./shared";
-import { fillTtml } from "@modules/lyrics/providers/ttmlUtils";
+import { fillTtml } from "@modules/lyrics/providers/ttmlSource";
 
 const JWT_RENEWAL_THRESHOLD_SECONDS = 60 * 60;
 
@@ -481,7 +480,7 @@ async function processStreamData(event: string, data: any, params: ProviderParam
       }
 
       if (results.plain) {
-        const lyrics = parsePlainLyrics(results.plain);
+        const lyrics = PlainParser.parse(results.plain);
         sourceMap["lrclib-plain"].lyricSourceResult = {
           lyrics,
           source: "LRCLib",
@@ -543,7 +542,7 @@ async function processStreamData(event: string, data: any, params: ProviderParam
         } catch (_e) {
           // Not double-encoded, use as is
         }
-        await fillTtml(ttml, params);
+        fillTtml(ttml, params);
         // fillTtml marks filled and updates sourceMap directly
         resolveWaiter(params, "bLyrics-synced");
         resolveWaiter(params, "bLyrics-richsynced");
@@ -553,7 +552,7 @@ async function processStreamData(event: string, data: any, params: ProviderParam
     // Binimum (TTML via BiniLyrics)
     if (provider === "binimum") {
       if (results.lyrics) {
-        await fillTtml(results.lyrics, params, {
+        fillTtml(results.lyrics, params, {
           richsyncKey: "binimum-richsynced",
           syncedKey: "binimum-synced",
           source: "BiniLyrics",
